+++ src/test/kotlin/me/bossm0n5t3r/leetcode/kthSmallestElementInASortedMatrix/KthSmallestElementInASortedMatrixTest.kt
@@ -1,10 +1,10 @@
-package kthSmallestElementInASortedMatrix
+package me.bossm0n5t3r.leetcode.kthSmallestElementInASortedMatrix
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 class KthSmallestElementInASortedMatrixTest {
-    private val kthSmallestElementInASortedMatrix = KthSmallestElementInASortedMatrix()
+    private val kthSmallestElementInASortedMatrix = KthSmallestElementInASortedMatrix.Solution()
 
     data class KthSmallestElementInASortedMatrixTestData(
         val matrix: Array<IntArray>,
